refactor(position):이동 관련 메소드를 MovePosition으로 이동
